requirements.txt
pandas==2.2.2
scikit-learn==1.5.0
joblib==1.4.2
google-cloud-storage==2.17.0
google-cloud-aiplatform==1.63.0
